+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Unreleased
 
 - Add `chrono::Duration` custom scalar. [#689](https://github.com/async-graphql/async-graphql/pull/689)
+- Implement `From<Option<Option<T>>>` for `MaybeUndefined<T>`.
+- Add `MaybeUndefined::as_opt_ref` and `MaybeUndefined::as_opt_deref` methods.
 
 ## [2.11.0] 2021-11-03